requirements.txt
alabaster 
anyio 
app==0.0.1
appdirs==1.4.4
argh==0.26.2
argon2-cffi 
asn1crypto 
astroid 
astropy 
async-generator 
atomicwrites==1.4.0
attrs 
autopep8 
Babel 
backcall 
backports.functools-lru-cache 
backports.shutil-get-terminal-size 
backports.tempfile 
backports.weakref==1.0.post1
bcrypt 
beautifulsoup4 
bitarray 
bkcharts==0.2
black==19.10b0
bleach 
bokeh 
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.12.5
cffi 
chardet 
Flask 
Jinja2 
matplotlib 
numpy 
opencv-python==4.5.2.54
pandas 
Pillow 
scikit-image==0.18.1
scikit-learn 
scipy 
torch==1.9.0
torchvision==0.10.0
tqdm 
gunicorn